pyproject.toml
[project]
name = "dreammaker-hft"
version = "0.1.0"
description = "High-Frequency Trading System with Advanced ML and Microstructure Analysis"
authors = [
    {name = "DreamMaker Team", email = "team@dreammaker.ai"}
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "numba>=0.58.0",
    "pyarrow>=14.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "joblib>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",
    "aiofiles>=23.0.0",
    "orjson>=3.9.0",
    "msgpack>=1.0.0",
    "statsmodels>=0.14.0",
    "arch>=6.2.0",
    "pywavelets>=1.4.0",
    "ipython>=8.37.0",
    "jupyter>=1.1.1",
    "notebook>=7.4.5",
    "pymt5linux",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.15.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipdb>=0.13.0",
]
windows = [
    "MetaTrader5>=5.0.45",
]

[project.scripts]
dreammaker = "src.main:cli"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "C90", "B", "S", "T10", "DJ", "RET", "SIM", "ARG", "PL"]
ignore = ["E501", "S101", "PLR0913"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
branch = true
source = ["src"]
